refactor(category): move categories into crud folder
