+++ FinEdgeFrontend/src/Components/CategoriesPage/CRUD/IncomeCategories.tsx
@@ -1,19 +1,19 @@
 import { Card, CardBody, CardFooter, CardHeader } from "@progress/kendo-react-layout";
-import usesGetExpenditureCategories from "../../Hooks/Categories/useGetExpenditureCategories";
-import CategoryCardBody from "./Cards/CategoryCardBody";
-import CategoryCardHeader from "./Cards/CategoryCardHeader";
-import CategoryCardFooter from "./Cards/CategoryCardFooter";
-import CategoriesDataCard from "./Cards/CategoriesDataCard";
+import usesGetIncomeCategories from "../../../Hooks/Categories/useGetIncomeCategories";
+import CategoryCardBody from "../Cards/CategoryCardBody";
+import CategoryCardHeader from "../Cards/CategoryCardHeader";
+import CategoryCardFooter from "../Cards/CategoryCardFooter";
+import CategoriesDataCard from "../Cards/CategoriesDataCard";
 
-const ExpenditureCategories = () => {
-  const { data, isLoading, isError, error } = usesGetExpenditureCategories();
+const IncomeCategories = () => {
+  const { data, isLoading, isError, error } = usesGetIncomeCategories();
 
   if (isLoading) return <p>Loading...</p>;
   if (isError) return <p>Error: {error!.message}</p>;
 
   return (
     <>
-      <h2 className="text-md font-semibold text-gray-600 mb-4 mt-4">Expenditure Categories</h2>
+      <h2 className="text-md font-semibold text-gray-600 mb-4 mt-4">Income Categories</h2>
       <div className="flex flex-col lg:flex-row gap-4">
         <div className="flex-grow">
           <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-4">
@@ -38,4 +38,4 @@ const ExpenditureCategories = () => {
   );
 };
 
-export default ExpenditureCategories;
+export default IncomeCategories;
